refactor: split types.go into domain-specific model files (#795)
